Fixed mmap usage for x86 Linux systems.
- huge pages are currently only used on Tilera systems (REMEMBER)
- Linux systems do not seem to map upwards like OS X (not only on Tilera)
- changed a few size related types from off_t to size_t, seems to be more consistent
- use _FILE_OFFSET_BITS = 64 to allow ftruncate to work with requests larger than 2GB-1 (REMEMBER, might be problematic on OS X, Tilera)

+++ vm/src/platform/thread_memory_semantics.cpp
@@ -108,7 +108,7 @@ u_int64 Thread_Memory_Semantics::my_rank_mask() {
 
 char  Thread_Memory_Semantics::mmap_filename[BUFSIZ] = { 0 };
 
-char* Thread_Memory_Semantics::map_heap_memory(off_t total_size,
+char* Thread_Memory_Semantics::map_heap_memory(size_t total_size,
                                                size_t bytes_to_map,
                                                size_t page_size_used_in_heap_arg,
                                                void* where,
